core: the 4,294,967,295-instruction ceiling for one run is gone
`ap_machine_run` took an `unsigned` limit and `ap_machine_run_t::executed`
counted in one, so a single run could not exceed 2^32-1 instructions. The
frontend refused a larger bound rather than wrapping it -- the right call,
and its comment said outright that "widening that is a core change with its
own verification and is not smuggled in here". This is that change.

It had stopped being an inconvenience. The DS5500 install needs ~15 G
instructions in one operation, with no oracle to borrow (`dn5500` is
MACHINE_NOT_WORKING) and no chaining available, because INVOL is re-entrant
one option at a time and a restore is not. The ceiling was that item's
blocker by name.

Both are uint64_t now, and -Wshorten-64-to-32 -Werror found every narrowing
rather than leaving one to surface later: two in the frontend, one in
ap_probe, and both boot entry points. ap_probe keeps its 32-bit field on
purpose -- ap_machine_run stops at the probe's own `unsigned` limit so that
narrowing provably cannot lose a bit, and widening it would rewrite every
golden for a range no probe can reach.

The frontend's refusal stays, narrowed to what it is for: strtoull saturates
at ULLONG_MAX, so a bound that is not a number, or is one this core cannot
represent, is still rejected rather than silently becoming a different bound.
That is the failure the original guard was written about.

Verification: machine_suite 63 -> 65. The discriminating test costs one
instruction rather than four billion: 0x100000000 is exactly 2^32, so a limit
narrowed to `unsigned` would be zero, and a program of one MOVEQ followed by
STOP separates honoured from truncated. The second asserts the count is the
same width as the bound. check_frontend_flags 25 -> 26, its ceiling check
replaced by one that a bound past 2^32 is accepted -- witnessed by the run
falling through to "cannot run dn3500 yet" rather than to a refusal -- plus
one that a non-numeric bound is still refused. ctest 145/145 both presets,
identity 6DF967A63D3D4DA9 unmoved.

+++ src/core/machine/ap_machine.c
@@ -920,10 +920,12 @@ static void machine_advance_devices(ap_machine_t *machine) {
   machine->last_line_f_exceptions = line_f;
 }
 
-ap_machine_run_t ap_machine_run(ap_machine_t *machine, unsigned limit) {
+ap_machine_run_t ap_machine_run(ap_machine_t *machine, uint64_t limit) {
   ap_machine_run_t out = {.status = AP_M68030_STEP_EXECUTED};
 
-  for (unsigned i = 0; i < limit; i++) {
+  /* 64-bit, so a run is bounded by what the caller asked for and not by the
+   * width of this counter. See `ap_machine_run_t::executed`. */
+  for (uint64_t i = 0; i < limit; i++) {
     const uint64_t before = machine->cpu.clocks;
     machine->instruction_start_clocks = before;
     if (machine->board != NULL) {
